feat(audio): add remainder fx
A Remainder effect is simple - it attaches a silent tail to an audio. It
is originally used to make reverb fade out fully audio-bounds to avoid
unpleasant effect when reverb is still fading out, but audio is out of
its length, and reverb's tail is hard-cut into a silence instantly. Yet,
it might be useful for some more effects in the future (e.g. delay).