fix(audio): add null check for audio object
The application is slightly (and will be pretty) multithreaded. In order
to avoid concurrency problems during memory management (when a memory
chunk gets cleaned up, but memory adress is still on the
`Allocation::ptr`, but at this moment `->raw()` is called). It is not
the solution, yet it's a step forward towards multithreading in Remixer.
There should be a smarter system than that though.