impr(memory)!: move from allocations to units
To provide a crystal-clean ownership system, allocations were replaced
with units. A unit is a origin-tolerant pointer wrapper that utilizes
stack cleaning to manage memory. Unlike stack variables, units are
intelligent enough to keep memory for other units if particular unit is
just a borrowed handler of the memory pointer (a.k.a. a "viewer"), or to
free memory and nullify its viewers if its an owner of the memory
pointer (a.k.a. a holder). Units are also thread-safe, yet there must be
some testing to ensure there are no deadlocks (which is unlikely by
using recursive mutexes), and are actually thread-safe (no racing
conditions). The whole audio system was rebuilt to use units.